action count is now in its own file
Count and Exists use the same underlying logic, which is why they used to be in the same file
That's confusing for a user so they are now in separate files.

Their tests are still both in `test_count` because they use the same fixtures and the project
is not currently set up to share fixtures accross modules.